Warn if CMS page is being translated
Show typical warning at top of block edit page.
Works for all pages with the same identifier.
Only applies if one of the active stores covers a target language.
If "All Store Views" then it will certainly import as a new page,
the current one won't be altered.
If multiple views but only some are target languages it is trickier,
this page might be split into multiple on import or might not.