Export non-ANSI characters as Unicode
Characters are no longer encoded as HTML entities.
This was causing double encoding.
Such characters could not be single encoded as XML entities are much
more restricted.
That leaves the option of decoding HTML back into Unicode or re-encoding
as numerical entities.
Since UTF-8 is so commonly understood in browsers and translation
software it makes sense to use the former, more compact approach.